requirements.txt
os
tkinter
easygui
shutil
argparse
collections
opencv-python
numpy
imagehash
Pillow